Dependency Error
React Hook useEffect has a missing dependency: 'fetchData', Either include
it or remove the dependency array - react-hooks/exhaustive-deps

This is because React detects that we are using an external value in the
hook that may change, and it assumes that we might want to run the hook
every time that value changes.

This is not our case, we only want to run the hook after the initial
render as in class lifecycle method (componentDidMount()) so we leave the
dependency array empty, which makes this warning not applicable for us.

More Info here:
https://reactjs.org/docs/hooks-effect.html#tip-optimizing-performance-by-skipping-effects